pyproject.toml
[project]
name = "crowthon"
version = "0.1.0"
description = "A powerful pluggable userbot on Telegram"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "telethon>=1.40.0",
]
